+++ packages/web/src/app/doc-pages/other-pages/color/color-token-table/colors.ts
@@ -10,6 +10,7 @@ import {
   DarkThemeColorName,
   LightThemeColorName,
   getThemeColor,
+  AssortedLabels,
 } from '@elvia/elvis-colors';
 
 /**
@@ -85,13 +86,19 @@ export const textColorsDefault = [
   },
   {
     token: 'text-3',
-    role: 'Placeholder text',
-    light: { hex: getColor('text-3', 'light'), label: 'Grey-70' },
-    dark: { hex: getColor('text-3', 'dark'), label: 'Dark-grey-20' },
+    role: 'Light text on dark grey background',
+    light: { hex: getColor('text-3', 'light'), label: 'White' },
+    dark: { hex: getColor('text-3', 'dark'), label: 'Dark-white' },
+    links: [{ title: 'Tooltip', path: '/components/tooltip' }],
+  },
+  {
+    token: 'text-4',
+    role: 'Text for dark backgrounds switching to light in dark theme',
+    light: { hex: getColor('text-4', 'light'), label: 'White' },
+    dark: { hex: getColor('text-4', 'dark'), label: 'Dark-black' },
     links: [
-      { title: 'Text field', path: '/components/input' },
-      { title: 'Date picker', path: '/components/datepicker' },
-      { title: 'Time picker', path: '/components/timepicker' },
+      { title: 'Button', path: '/components/button' },
+      { title: 'Badge', path: '/components/badge' },
     ],
   },
 ] as const satisfies TableColorArray<TextLabels>;
@@ -117,6 +124,18 @@ export const textColorsState = [
     links: [
       { title: 'Primary button', path: '/components/button' },
       { title: 'Danger button', path: '/components/button' },
+    ],
+  },
+  {
+    token: 'text-placeholder-1',
+    role: 'Placeholder text',
+    light: { hex: getColor('text-placeholder-1', 'light'), label: 'Grey-70' },
+    dark: { hex: getColor('text-placeholder-1', 'dark'), label: 'Dark-grey-20' },
+    links: [
+      { title: 'Text field', path: '/components/input' },
+      { title: 'Date picker', path: '/components/datepicker' },
+      { title: 'Time picker', path: '/components/timepicker' },
+      { title: 'Slider', path: '/components/slider' },
     ],
   },
 ] as const satisfies TableColorArray<TextLabels>;
@@ -152,7 +171,7 @@ export const backgroundColorsElement = [
   {
     token: 'background-element-2',
     role: 'Element background color option 2',
-    light: { hex: getColor('background-element-2', 'light'), label: 'Grey-05' },
+    light: { hex: getColor('background-element-2', 'light'), label: 'Grey-02' },
     dark: { hex: getColor('background-element-2', 'dark'), label: 'Dark-grey-70' },
     links: [{ title: 'Table (zebra row)', path: '/components/table' }],
   },
@@ -175,6 +194,16 @@ export const backgroundColorsElement = [
     light: { hex: getColor('background-element-5', 'light'), label: 'White' },
     dark: { hex: getColor('background-element-5', 'dark'), label: 'Dark-grey-70' },
     links: [{ title: 'Box', path: '/components/box' }],
+  },
+  {
+    token: 'background-element-6',
+    role: 'Element background color option 6',
+    light: { hex: getColor('background-element-6', 'light'), label: 'Black' },
+    dark: { hex: getColor('background-element-6', 'dark'), label: 'Dark-white' },
+    links: [
+      { title: 'Button', path: '/components/button' },
+      { title: 'Badge', path: '/components/badge' },
+    ],
   },
 ] as const satisfies TableColorArray<BackgroundLabels>;
 
@@ -476,3 +505,30 @@ export const iconColors = [
     dark: { hex: 'dark-signal', label: 'Dark-signal' },
   },
 ] as const satisfies TableColorArray<IconLabels | 'icon-[signal-type]'>;
+
+export const assortedColors = [
+  {
+    token: 'static-black',
+    role: 'Stays black in both themes',
+    light: { hex: getColor('static-black', 'light'), label: 'Black' },
+    dark: { hex: getColor('static-black', 'dark'), label: 'Dark-black' },
+  },
+  {
+    token: 'static-white',
+    role: 'Stays white in both themes',
+    light: { hex: getColor('static-white', 'light'), label: 'White' },
+    dark: { hex: getColor('static-white', 'dark'), label: 'Dark-white' },
+  },
+  {
+    token: 'brand-accent',
+    role: 'Brand color that can be used for accents like bullet points in lists',
+    light: { hex: getColor('brand-accent', 'light'), label: 'Green' },
+    dark: { hex: getColor('brand-accent', 'dark'), label: 'Dark-green' },
+  },
+  {
+    token: 'focus-outline',
+    role: 'Color only used for focus outline. Should not be used elsewhere.',
+    light: { hex: getColor('focus-outline', 'light'), label: 'Focus-outline' },
+    dark: { hex: getColor('focus-outline', 'dark'), label: 'Dark-focus-outline' },
+  },
+] as const satisfies TableColorArray<AssortedLabels>;